Moved fee calculation into separate source file
