+++ libraries/protocol/fee_schedule.cpp
@@ -25,7 +25,6 @@
 #include <graphene/protocol/fee_schedule.hpp>
 
 #include <fc/io/raw.hpp>
-#include <fc/uint128.hpp>
 
 #define MAX_FEE_STABILIZATION_ITERATION 4
 
@@ -45,28 +44,6 @@ namespace graphene { namespace protocol {
       }
       return result;
    }
-
-   struct calc_fee_visitor
-   {
-      typedef uint64_t result_type;
-
-      const fee_schedule& param;
-      const int current_op;
-      calc_fee_visitor( const fee_schedule& p, const operation& op ):param(p),current_op(op.which()){}
-
-      template<typename OpType>
-      result_type operator()( const OpType& op )const
-      {
-         try {
-            return op.calculate_fee( param.get<OpType>() ).value;
-         } catch (fc::assert_exception& e) {
-             fee_parameters params; params.set_which(current_op);
-             auto itr = param.parameters.find(params);
-             if( itr != param.parameters.end() ) params = *itr;
-             return op.calculate_fee( params.get<typename OpType::fee_parameters_type>() ).value;
-         }
-      }
-   };
 
    struct set_fee_visitor
    {
@@ -99,25 +76,6 @@ namespace graphene { namespace protocol {
       for( fee_parameters& i : parameters )
          i.visit( zero_fee_visitor() );
       this->scale = 0;
-   }
-
-   asset fee_schedule::calculate_fee( const operation& op )const
-   {
-      uint64_t required_fee = op.visit( calc_fee_visitor( *this, op ) );
-      if( scale != GRAPHENE_100_PERCENT )
-      {
-         auto scaled = fc::uint128_t(required_fee) * scale;
-         scaled /= GRAPHENE_100_PERCENT;
-         FC_ASSERT( scaled <= GRAPHENE_MAX_SHARE_SUPPLY,
-                    "Required fee after scaling would exceed maximum possible supply" );
-         required_fee = static_cast<uint64_t>(scaled);
-      }
-      return asset( required_fee );
-   }
-
-   asset fee_schedule::calculate_fee( const operation& op, const price& core_exchange_rate )const
-   {
-      return calculate_fee( op ).multiply_and_round_up( core_exchange_rate );
    }
 
    asset fee_schedule::set_fee( operation& op, const price& core_exchange_rate )const
@@ -168,5 +126,3 @@ namespace graphene { namespace protocol {
    }
 
 } } // graphene::protocol
-
-GRAPHENE_IMPLEMENT_EXTERNAL_SERIALIZATION( graphene::protocol::fee_schedule )
